Advertise and scan for the service over BLE
This change implements BLE advertisements and scanning, but in a
slightly inefficent way: instead of using ScanFilters to let the radio
check if another instance of this app produced an advertisement, I check
app-side in the ScanCallback. This is because ScanFilters don't seem to
do anything unless another app is actively doing a scan?

It may be worth revisiting ScanFilters if this implementation uses too
much power.