Automatically connect to other devices in range
This change implements automatic connections to other devices. Because
the BluetoothDevice we receive in a BLE scan is for a bogus device
(Android BLE advertisements use a random MAC address), we include the
address (sometimes obtained using reflection because Android Marshmallow
and up block getting it normally) in the advertisement. Once we detect a
nearby device, we connect to it using Bluetooth insecure sockets (to
avoid pairing) and use a simple handshake to verify that the connection
works.