WIP: Implement UnknownMessage I/O methods
This change provides preliminary implementations of UnknownMessage's I/O
methods, specifically (de)serialization and creation of a new message from a
byte array (which is automatically signed using hashcash).

When creating a message from a byte array, it is always padded with random data
so that the stored payload is 240 bytes long to improve anonymity - if the byte
array passed in is encrypted, observers will be unable to tell how long the
actual data is because it all looks like random data. The intent is that users
of UnknownMessage will store their own length in the payload.

The change renames UnknownMessage.data to UnknownMessage.payload for
readability and updates its only user in the test view to use the updated APIs.

TODOs:
* Tune the payload length
* Determine message expiry from declared number of zero bytes
* Tests (probably integration tests because we will need a real database to
  test against)