Clean up logcat output when signing
Previously, we were logging in isValid() each time a message was invalid
because it didn't have enough zero bits in its hash. Because this
function was called as part of the signing loop, we were spamming the
logcat with millions of lines each time we made a new message. This
change removes these log calls.

This change also logs how long it took a message to sign.