WIP: Add unit tests for UnknownMessage I/O
This change adds a unit test that validates that
UnknownMessage.createAndSignAsync can be called without error.

For now, I've manually validated using a breakpoint that signing runs
and that save is called, but the test does not yet check the database
after the message is saved.

I changed UnknownMessage.create* to return the DBFlow Transaction so
that we can wait for it to finish in the test.

Remaining work:
* Make sure that the message exists in the database
* Retrieve the message and check that it is signed properly
* Validate the payload of the message