Implement bloom filter message vector generation
This change implements generating a bloom filter message vector. Much of
the implementation is borrowed from this guide:

http://blog.michaelschmatz.com/2016/04/11/how-to-write-a-bloom-filter-cpp/

I imported a Java implementation of the MurmurHash3 hash here:

http://yonik.com/murmurhash3-for-java/

This change introduces a new table, BloomFilter, that maps an
UnknownMessage to each of its hash values for sync, along with basic
tests for it. My implementation currently does not convert this
representation down to a single bloom filter bit vector until it is
explicitly requested.

NYI:

* Listing the UnknownMessage(s) that satisfy a set of hashes
* Listing the UnknownMessages that are *not* described by a given bit
  vector
* Using the above to determine what needs to be sent over during sync