Make the handshake test actually wait for handshake to finish
The previous test only appeared to run properly because calling
shutdown() on an ExecutorService doesn't actually block, meaning that
the test was never actually waiting for the I/O threads to return. This
change has handshakeAsync return a new IOFutures object that has a get()
method that waits for both send and receive threads to finish, which
both the handshake test and bidirectional sync now use to wait for
handshake to finish.