Implement and test message vector exchange
This change implements message vector exchange and tests it by
exchanging two fake message vectors and making sure they're the same on
the other side. While working on this, I noticed that BitSet's size is
determined solely by the highest bit that is set regardless of what is
passed into the constructor, so message vectors are now required to have
their last bit set so that they always have the same size.