BROKEN: Use RxJava in the BloomFilter implementation
This enables calculating the message vector (getMessageVectorAsync) without
loading all of the messages' hashes into memory at once(!), and allows doing so
asynchronously.

Unfortunately, this doesn't work both as a unit test and on physical devices
since the BiFunction callback is never called.

Maybe this is because of a bug in queryStreamResults?
https://github.com/Raizlabs/DBFlow/pull/1262