Don't establish more than one simultaneous connection to a device
Previously, BluetoothSyncService would *always* try to connect to a
device after it discovers it, regardless of whether it was already
connected to that device. BLE scans are continuous and continue to
return the same device until it leaves range, causing multiple
simultaneous connections between a pair of devices.

This change avoids this by keeping track of the connected devices.

Sync is still broken though - handshake never seems to complete.